+++ tests/integration/test_customer.py
@@ -308,7 +308,7 @@ class TestCustomer(unittest.TestCase):
         try:
             Customer.find("badid")
             self.assertTrue(False)
-        except NotFoundError, e:
+        except NotFoundError as e:
             self.assertEquals("customer with id badid not found", str(e))
 
     def test_update_with_valid_options(self):