+++ tests/unit/test_credit_card.py
@@ -6,28 +6,28 @@ class TestCreditCard(unittest.TestCase):
         try:
             CreditCard.create({"bad_key": "value"})
             self.assertTrue(False)
-        except KeyError, e:
+        except KeyError as e:
             self.assertEquals("'Invalid keys: bad_key'", str(e))
 
     def test_update_raises_exception_with_bad_keys(self):
         try:
             CreditCard.update("token", {"bad_key": "value"})
             self.assertTrue(False)
-        except KeyError, e:
+        except KeyError as e:
             self.assertEquals("'Invalid keys: bad_key'", str(e))
 
     def test_tr_data_for_create_raises_error_with_bad_keys(self):
         try:
             CreditCard.tr_data_for_create({"bad_key": "value"}, "http://example.com")
             self.assertTrue(False)
-        except KeyError, e:
+        except KeyError as e:
             self.assertEquals("'Invalid keys: bad_key'", str(e))
 
     def test_tr_data_for_update_raises_error_with_bad_keys(self):
         try:
             CreditCard.tr_data_for_update({"bad_key": "value"}, "http://example.com")
             self.assertTrue(False)
-        except KeyError, e:
+        except KeyError as e:
             self.assertEquals("'Invalid keys: bad_key'", str(e))
 
     def test_transparent_redirect_create_url(self):
@@ -84,6 +84,6 @@ class TestCreditCard(unittest.TestCase):
         try:
             CreditCard.find(" ")
             self.assertTrue(False)
-        except NotFoundError, e:
+        except NotFoundError as e:
             self.assertTrue(True)